pyproject.toml
[tool.poetry]
name = "pytry"
version = "0.1.0"
description = ""
authors = ["Jan Ouwens"]

[tool.poetry.dependencies]
python = "^3.7"
python-dateutil = "^2.8.1"
BeautifulSoup4 = "^4.8.2"

[tool.poetry.dev-dependencies]
pytest = "^4.6"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
